profile page : user sees his artworks

+++ images/Frontend/src/routes/Profile.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import User from "../classes/User";
+import Artwork from "../classes/Artwork";
 import "./styles/Profile.css";
 
 const ProfilePage = () => {
@@ -10,7 +11,28 @@ const ProfilePage = () => {
   const [newEmail, setNewEmail] = useState("");
   const [editMode, setEditMode] = useState(false);
   const [error, setError] = useState("");
+  const [confirmDelete, setConfirmDelete] = useState(false);
   const [successMessage, setSuccessMessage] = useState("");
+  const [artworks, setArtworks] = useState([]);
+
+  const fetchUserArtworks = async () => {
+    try {
+      const userId = localStorage.getItem("user_ID");
+
+      const response = await fetch(`http://localhost/artworks/user/${userId}`);
+      const data = await response.json();
+
+      if (data.message === "No artworks found for the specified user.") {
+        setArtworks([]);
+      } else {
+        const artworks = data.map((artworkData) => new Artwork(artworkData));
+        console.log("Mapped Artworks:", artworks);
+        setArtworks(artworks);
+      }
+    } catch (error) {
+      console.error("Error fetching artworks:", error);
+    }
+  };
 
   useEffect(() => {
     // Retrieve user information from local storage
@@ -25,6 +47,8 @@ const ProfilePage = () => {
       // Set the user state to use in the component
       setUser(profile);
     }
+
+    fetchUserArtworks();
   }, []);
 
   const handleUpdate = () => {
@@ -70,6 +94,11 @@ const ProfilePage = () => {
   };
 
   const handleDeleteProfile = () => {
+    if (!confirmDelete) {
+      setConfirmDelete(true);
+      return;
+    }
+
     fetch(`http://localhost/users/${user.id}`, {
       method: "DELETE",
       headers: {
@@ -96,55 +125,95 @@ const ProfilePage = () => {
 
   return (
     <div>
-      <h1>Profile Page</h1>
       {user ? (
         <div className="profile">
-          <h2>Welcome, {user.name}!</h2>
-          <h3>Here are your profile details:</h3>
-          <div className="email">
-            <h4>Edit your email and password?</h4>
-            <p>Current email: {user.email}</p>
+          <h1>Welcome, {user.name}!</h1>
+          <div className="row ">
             <div>
-              {editMode && (
-                <>
-                  <label>New Email:</label>
-                  <input
-                    type="email"
-                    value={newEmail}
-                    onChange={(e) => setNewEmail(e.target.value)}
-                  />
-                </>
+              <h3>Here are your profile details:</h3>
+              <div className="email">
+                <h4>Edit your email and password?</h4>
+                <p>Current email: {user.email}</p>
+                <div>
+                  {editMode && (
+                    <>
+                      <label>New Email:</label>
+                      <input
+                        type="email"
+                        value={newEmail}
+                        onChange={(e) => setNewEmail(e.target.value)}
+                      />
+                    </>
+                  )}
+                </div>
+              </div>
+              <div className="password">
+                {editMode && (
+                  <div>
+                    <label>New Password:</label>
+                    <input
+                      type="password"
+                      value={newPassword}
+                      onChange={(e) => setNewPassword(e.target.value)}
+                    />
+                    <label>Confirm new Password:</label>
+                    <input
+                      type="password"
+                      value={confirmPassword}
+                      onChange={(e) => setConfirmPassword(e.target.value)}
+                    />
+                  </div>
+                )}
+              </div>
+              {error && <p style={{ color: "red" }}>{error}</p>}
+              {successMessage && (
+                <p style={{ color: "green" }}>{successMessage}</p>
               )}
+              <div className="buttons">
+                {editMode ? (
+                  <>
+                    <button onClick={handleUpdate}>Confirm</button>
+                    <button onClick={() => setEditMode(false)}>Cancel</button>
+                  </>
+                ) : (
+                  <>
+                    <button onClick={() => setEditMode(true)}>Edit</button>
+
+                    <button onClick={() => { setConfirmDelete(!confirmDelete);}}>
+
+                      {confirmDelete ? "Cancel" : "Delete Profile"}
+                      
+                    </button>
+                    {confirmDelete && (
+                      <button onClick={handleDeleteProfile}>
+                        Confirm Deletion
+                      </button>
+                    )}
+                  </>
+                )}
+              </div>
+            </div>
+            <div className="artworks">
+              <div className="artworks-container">
+                <div className="artworks">
+                  <h3>Your Artworks:</h3>
+                  {artworks.length > 0 ? (
+                    <ul>
+                      {artworks.map((artwork) => (
+                        <div className="artwork-card" key={artwork.id}>
+                          <img src={artwork.image} alt={artwork.name} />
+                          <p>Name: {artwork.name}</p>
+                          <p>Description: {artwork.description}</p>
+                        </div>
+                      ))}
+                    </ul>
+                  ) : (
+                    <p>No artworks found for the specified user.</p>
+                  )}
+                </div>
+              </div>
             </div>
           </div>
-          <div className="password">
-            {editMode && (
-              <div>
-                <label>New Password:</label>
-                <input
-                  type="password"
-                  value={newPassword}
-                  onChange={(e) => setNewPassword(e.target.value)}
-                />
-                <label>Confirm new Password:</label>
-                <input
-                  type="password"
-                  value={confirmPassword}
-                  onChange={(e) => setConfirmPassword(e.target.value)}
-                />
-              </div>
-            )}
-          </div>
-          {error && <p style={{ color: "red" }}>{error}</p>}
-          {successMessage && <p style={{ color: "green" }}>{successMessage}</p>}
-          <div className="buttons">
-            {editMode ? (
-              <button onClick={handleUpdate}>Confirm</button>
-            ) : (
-              <button onClick={() => setEditMode(true)}>Edit</button>
-            )}
-          </div>
-          <button onClick={handleDeleteProfile}>Delete Profile</button>
         </div>
       ) : (
         <p>User not logged in.</p>
